Fold TcpMessageHandler into TcpTransport
