+++ core/src/main/java/org/elasticsearch/transport/TcpTransport.java
@@ -20,6 +20,7 @@ package org.elasticsearch.transport;
 
 import com.carrotsearch.hppc.IntHashSet;
 import com.carrotsearch.hppc.IntSet;
+import org.apache.lucene.util.IOUtils;
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.ExceptionsHelper;
 import org.elasticsearch.Version;
@@ -29,9 +30,13 @@ import org.elasticsearch.common.breaker.CircuitBreaker;
 import org.elasticsearch.common.bytes.BytesArray;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.component.AbstractLifecycleComponent;
+import org.elasticsearch.common.component.Lifecycle;
+import org.elasticsearch.common.compress.Compressor;
 import org.elasticsearch.common.compress.CompressorFactory;
+import org.elasticsearch.common.compress.NotCompressedException;
 import org.elasticsearch.common.io.ReleasableBytesStream;
 import org.elasticsearch.common.io.stream.BytesStreamOutput;
+import org.elasticsearch.common.io.stream.NamedWriteableAwareStreamInput;
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.io.stream.ReleasableBytesStreamOutput;
 import org.elasticsearch.common.io.stream.StreamInput;
@@ -52,7 +57,9 @@ import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.common.util.BigArrays;
 import org.elasticsearch.common.util.concurrent.AbstractLifecycleRunnable;
+import org.elasticsearch.common.util.concurrent.AbstractRunnable;
 import org.elasticsearch.common.util.concurrent.KeyedLock;
+import org.elasticsearch.common.util.concurrent.ThreadContext;
 import org.elasticsearch.indices.breaker.CircuitBreakerService;
 import org.elasticsearch.monitor.jvm.JvmInfo;
 import org.elasticsearch.rest.RestStatus;
@@ -532,11 +539,6 @@ public abstract class TcpTransport<Channel> extends AbstractLifecycleComponent<T
             result.put(name, mergedSettings);
         }
         return result;
-    }
-
-
-    public TcpMessageHandler newMessageChannelHandler() {
-        return new TcpMessageHandler(threadPool, this, transportServiceAdapter, namedWriteableRegistry, logger);
     }
 
     @Override
@@ -1141,5 +1143,209 @@ public abstract class TcpTransport<Channel> extends AbstractLifecycleComponent<T
     }
 
     protected abstract boolean isOpen(Channel channel);
+
+    public final void messageReceived(BytesReference reference, ChannelFactory channelFactory,
+                                      InetSocketAddress remoteAddress, int messageLengthBytes) throws IOException {
+        final int totalMessageSize = messageLengthBytes + TcpHeader.MARKER_BYTES_SIZE + TcpHeader.MESSAGE_LENGTH_SIZE;
+        transportServiceAdapter.received(totalMessageSize);
+        // we have additional bytes to read, outside of the header
+        boolean hasMessageBytesToRead = (totalMessageSize - TcpHeader.HEADER_SIZE) > 0;
+        StreamInput streamIn = reference.streamInput();
+        boolean success = false;
+        try (ThreadContext.StoredContext tCtx = threadPool.getThreadContext().stashContext()) {
+            long requestId = streamIn.readLong();
+            byte status = streamIn.readByte();
+            Version version = Version.fromId(streamIn.readInt());
+            if (TransportStatus.isCompress(status) && hasMessageBytesToRead && streamIn.available() > 0) {
+                Compressor compressor;
+                try {
+                    final int bytesConsumed = TcpHeader.REQUEST_ID_SIZE + TcpHeader.STATUS_SIZE + TcpHeader.VERSION_ID_SIZE;
+                    compressor = CompressorFactory.compressor(reference.slice(bytesConsumed, reference.length() - bytesConsumed));
+                } catch (NotCompressedException ex) {
+                    int maxToRead = Math.min(reference.length(), 10);
+                    StringBuilder sb = new StringBuilder("stream marked as compressed, but no compressor found, first [").append(maxToRead)
+                        .append("] content bytes out of [").append(reference.length())
+                        .append("] readable bytes with message size [").append(messageLengthBytes).append("] ").append("] are [");
+                    for (int i = 0; i < maxToRead; i++) {
+                        sb.append(reference.get(i)).append(",");
+                    }
+                    sb.append("]");
+                    throw new IllegalStateException(sb.toString());
+                }
+                streamIn = compressor.streamInput(streamIn);
+            }
+            if (version.onOrAfter(Version.CURRENT.minimumCompatibilityVersion()) == false || version.major != Version.CURRENT.major) {
+                throw new IllegalStateException("Received message from unsupported version: [" + version
+                    + "] minimal compatible version is: [" +Version.CURRENT.minimumCompatibilityVersion() + "]");
+            }
+            streamIn = new NamedWriteableAwareStreamInput(streamIn, namedWriteableRegistry);
+            streamIn.setVersion(version);
+            if (TransportStatus.isRequest(status)) {
+                threadPool.getThreadContext().readHeaders(streamIn);
+                handleRequest(channelFactory, streamIn, requestId, messageLengthBytes, version, remoteAddress);
+            } else {
+                final TransportResponseHandler<?> handler = transportServiceAdapter.onResponseReceived(requestId);
+                // ignore if its null, the adapter logs it
+                if (handler != null) {
+                    if (TransportStatus.isError(status)) {
+                        handlerResponseError(streamIn, handler);
+                    } else {
+                        handleResponse(remoteAddress, streamIn, handler);
+                    }
+                    // Check the entire message has been read
+                    final int nextByte = streamIn.read();
+                    // calling read() is useful to make sure the message is fully read, even if there is an EOS marker
+                    if (nextByte != -1) {
+                        throw new IllegalStateException("Message not fully read (response) for requestId [" + requestId + "], handler ["
+                            + handler + "], error [" + TransportStatus.isError(status) + "]; resetting");
+                    }
+                }
+            }
+            success = true;
+        } finally {
+            if (success) {
+                IOUtils.close(streamIn);
+            } else {
+                IOUtils.closeWhileHandlingException(streamIn);
+            }
+        }
+    }
+
+    protected void handleResponse(InetSocketAddress remoteAddress, final StreamInput stream, final TransportResponseHandler handler) {
+        final TransportResponse response = handler.newInstance();
+        response.remoteAddress(new InetSocketTransportAddress(remoteAddress));
+        try {
+            response.readFrom(stream);
+        } catch (Throwable e) {
+            handleException(handler, new TransportSerializationException(
+                "Failed to deserialize response of type [" + response.getClass().getName() + "]", e));
+            return;
+        }
+        threadPool.executor(handler.executor()).execute(new AbstractRunnable() {
+            @Override
+            public void onFailure(Throwable t) {
+                handleException(handler, new ResponseHandlerFailureTransportException(t));
+            }
+
+            @Override
+            protected void doRun() throws Exception {
+                handler.handleResponse(response);
+            }});
+
+    }
+
+    /**
+     * Executed for a received response error
+     */
+    private void handlerResponseError(StreamInput stream, final TransportResponseHandler handler) {
+        Throwable error;
+        try {
+            error = stream.readThrowable();
+        } catch (Throwable e) {
+            error = new TransportSerializationException("Failed to deserialize exception response from stream", e);
+        }
+        handleException(handler, error);
+    }
+
+    private void handleException(final TransportResponseHandler handler, Throwable error) {
+        if (!(error instanceof RemoteTransportException)) {
+            error = new RemoteTransportException(error.getMessage(), error);
+        }
+        final RemoteTransportException rtx = (RemoteTransportException) error;
+        threadPool.executor(handler.executor()).execute(() -> {
+            try {
+                handler.handleException(rtx);
+            } catch (Throwable e) {
+                logger.error("failed to handle exception response [{}]", e, handler);
+            }
+        });
+    }
+
+    protected String handleRequest(ChannelFactory channelFactory, final StreamInput stream, long requestId,
+                                   int messageLengthBytes, Version version, InetSocketAddress remoteAddress) throws IOException {
+        final String action = stream.readString();
+        transportServiceAdapter.onRequestReceived(requestId, action);
+        TransportChannel transportChannel = null;
+        try {
+            final RequestHandlerRegistry reg = transportServiceAdapter.getRequestHandler(action);
+            if (reg == null) {
+                throw new ActionNotFoundTransportException(action);
+            }
+            if (reg.canTripCircuitBreaker()) {
+                getInFlightRequestBreaker().addEstimateBytesAndMaybeBreak(messageLengthBytes, "<transport_request>");
+            } else {
+                getInFlightRequestBreaker().addWithoutBreaking(messageLengthBytes);
+            }
+            transportChannel = channelFactory.create(action, requestId, version, messageLengthBytes);
+            final TransportRequest request = reg.newRequest();
+            request.remoteAddress(new InetSocketTransportAddress(remoteAddress));
+            request.readFrom(stream);
+            // in case we throw an exception, i.e. when the limit is hit, we don't want to verify
+            validateRequest(stream, requestId, action);
+            threadPool.executor(reg.getExecutor()).execute(new RequestHandler(reg, request, transportChannel));
+        } catch (Throwable e) {
+            // the circuit breaker tripped
+            if (transportChannel == null) {
+                transportChannel = channelFactory.create(action, requestId, version, 0);
+            }
+            try {
+                transportChannel.sendResponse(e);
+            } catch (IOException e1) {
+                logger.warn("Failed to send error message back to client for action [{}]", e, action);
+                logger.warn("Actual Exception", e1);
+            }
+        }
+        return action;
+    }
+
+    // This template method is needed to inject custom error checking logic in tests.
+    protected void validateRequest(StreamInput stream, long requestId, String action) throws IOException {
+        final int nextByte = stream.read();
+        // calling read() is useful to make sure the message is fully read, even if there some kind of EOS marker
+        if (nextByte != -1) {
+            throw new IllegalStateException("Message not fully read (request) for requestId [" + requestId + "], action [" + action
+                + "], available [" + stream.available() + "]; resetting");
+        }
+    }
+
+    public interface ChannelFactory {
+        TransportChannel create(String actionName, long requestId, Version version, long reservedBytes);
+    }
+
+    class RequestHandler extends AbstractRunnable {
+        private final RequestHandlerRegistry reg;
+        private final TransportRequest request;
+        private final TransportChannel transportChannel;
+
+        public RequestHandler(RequestHandlerRegistry reg, TransportRequest request, TransportChannel transportChannel) {
+            this.reg = reg;
+            this.request = request;
+            this.transportChannel = transportChannel;
+        }
+
+        @SuppressWarnings({"unchecked"})
+        @Override
+        protected void doRun() throws Exception {
+            reg.processMessageReceived(request, transportChannel);
+        }
+
+        @Override
+        public boolean isForceExecution() {
+            return reg.isForceExecution();
+        }
+
+        @Override
+        public void onFailure(Throwable e) {
+            if (lifecycleState() == Lifecycle.State.STARTED) {
+                // we can only send a response transport is started....
+                try {
+                    transportChannel.sendResponse(e);
+                } catch (Throwable e1) {
+                    logger.warn("Failed to send error message back to client for action [{}]", e1, reg.getAction());
+                    logger.warn("Actual Exception", e);
+                }
+            }
+        }
+    }
 
 }
