+++ core/src/main/java/org/elasticsearch/transport/netty/NettyMessageChannelHandler.java
@@ -20,8 +20,8 @@
 package org.elasticsearch.transport.netty;
 
 import org.elasticsearch.common.bytes.BytesReference;
-import org.elasticsearch.transport.TcpMessageHandler;
 import org.elasticsearch.transport.TcpHeader;
+import org.elasticsearch.transport.TcpTransport;
 import org.elasticsearch.transport.TcpTransportChannel;
 import org.elasticsearch.transport.TransportServiceAdapter;
 import org.elasticsearch.transport.Transports;
@@ -43,10 +43,8 @@ class NettyMessageChannelHandler extends SimpleChannelUpstreamHandler {
     protected final TransportServiceAdapter transportServiceAdapter;
     protected final NettyTransport transport;
     protected final String profileName;
-    private final TcpMessageHandler messageChannelHandler;
 
-    NettyMessageChannelHandler(NettyTransport transport, String profileName, TcpMessageHandler handler) {
-        messageChannelHandler = handler;
+    NettyMessageChannelHandler(NettyTransport transport, String profileName) {
         this.transportServiceAdapter = transport.transportServiceAdapter();
         this.transport = transport;
         this.profileName = profileName;
@@ -69,14 +67,14 @@ class NettyMessageChannelHandler extends SimpleChannelUpstreamHandler {
         final ChannelBuffer buffer = (ChannelBuffer) m;
         final int remainingMessageSize = buffer.getInt(buffer.readerIndex() - TcpHeader.MESSAGE_LENGTH_SIZE);
         final int expectedReaderIndex = buffer.readerIndex() + remainingMessageSize;
-        TcpMessageHandler.ChannelFactory channelFactory = (action, requestId, version, reservedBytes) ->
+        TcpTransport.ChannelFactory channelFactory = (action, requestId, version, reservedBytes) ->
             new TcpTransportChannel<>(transport, ctx.getChannel(), "netty", action, requestId, version, profileName, reservedBytes);
         InetSocketAddress remoteAddress = (InetSocketAddress) ctx.getChannel().getRemoteAddress();
         try {
             // netty always copies a buffer, either in NioWorker in its read handler, where it copies to a fresh
             // buffer, or in the cumulation buffer, which is cleaned each time so it could be bigger than the actual size
             BytesReference reference = NettyUtils.toBytesReference(buffer, remainingMessageSize);
-            messageChannelHandler.messageReceived(reference, channelFactory, remoteAddress, remainingMessageSize);
+            transport.messageReceived(reference, channelFactory, remoteAddress, remainingMessageSize);
         } finally {
             // Set the expected position of the buffer, no matter what happened
             buffer.readerIndex(expectedReaderIndex);
